requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pdfplumber
python-docx